perf: oas dereferencing when we load operations, not docs
